+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * `replication_primary_canary_age_ms`
   * `enterprise`
 * Add missing `num_uses` field to `AuthData`
+* Add `mount_type` attribute to common response model
 
 ### Dependencies
 * Updated Jackson to 2.17.1